+++ src/cli/completions.rs
@@ -1,28 +1,15 @@
 use crate::cli::Arguments;
-use clap::{Command, CommandFactory};
-use clap_complete::{Generator, Shell, generate};
+use clap::CommandFactory;
+use clap_complete::generate;
 use std::io;
 use std::process::exit;
 
 pub fn generate_shell_completions_and_exit_or_continue(cli: &Arguments) {
-    let mut cmd = Arguments::command();
+    if let Some(shell) = cli.completions {
+        let mut cmd = Arguments::command();
+        let name = cmd.get_name().to_string();
 
-    if cli.generate_bash_completions {
-        print_completions(Shell::Bash, &mut cmd);
+        generate(shell, &mut cmd, name, &mut io::stdout());
         exit(0);
     }
-
-    if cli.generate_fish_completions {
-        print_completions(Shell::Fish, &mut cmd);
-        exit(0);
-    }
-
-    if cli.generate_zsh_completions {
-        print_completions(Shell::Zsh, &mut cmd);
-        exit(0);
-    }
-}
-
-fn print_completions<G: Generator>(generator: G, cmd: &mut Command) {
-    generate(generator, cmd, cmd.get_name().to_string(), &mut io::stdout());
 }